requirements.txt
django==1.6
djangorestframework==2.3.13
requests
python-social-auth
django-celery
djangorestframework